(BSR)[API] chore: Move `recredit_underage_users()` to `core.finance` package
